refactor(clip): 애플리케이션 DTO 구조 정리
