+++ src/clips/application/dtos/clip-cursor-page-output.dto.ts
@@ -1,0 +1,12 @@
+import type { ClipListItem } from '../../domain/clip.types';
+
+export type ClipCursorPageOutput = {
+  items: ClipListItem[];
+  nextCursor: string | null;
+};
+
+export type FavoriteClipsPageOutput = {
+  items: ClipListItem[];
+  hasMore: boolean;
+  nextCursor: string | null;
+};
