+++ src/clips/application/dtos/create-clip-input.dto.ts
@@ -1,0 +1,4 @@
+export type CreateClipInput = {
+  folderId: string;
+  text?: string;
+};
