+++ src/clips/application/dtos/like-clip-output.dto.ts
@@ -1,0 +1,3 @@
+export type LikeClipOutput = {
+  likeByMe: boolean;
+};
