+++ src/clips/application/dtos/list-folder-clips-input.dto.ts
@@ -1,0 +1,8 @@
+import type { ClipTypeFilter } from '../../domain/clips.repository';
+
+export type ListFolderClipsInput = {
+  folderId: string;
+  cursor?: string;
+  type: ClipTypeFilter;
+  q?: string;
+};
