+++ src/clips/application/dtos/list-recent-clips-input.dto.ts
@@ -1,0 +1,7 @@
+import type { ClipTypeFilter } from '../../domain/clips.repository';
+
+export type ListRecentClipsInput = {
+  cursor?: string;
+  type: ClipTypeFilter;
+  q?: string;
+};
