+++ src/clips/application/dtos/recent-viewed-clips-output.dto.ts
@@ -1,0 +1,5 @@
+import type { ClipListItem } from '../../domain/clip.types';
+
+export type RecentViewedClipsOutput = {
+  items: ClipListItem[];
+};
