+++ src/clips/application/dtos/update-clip-input.dto.ts
@@ -1,0 +1,5 @@
+export type UpdateClipInput = {
+  clipId: string;
+  folderId?: string;
+  text?: string;
+};
