+++ src/clips/application/usecases/unlike-clip.usecase.ts
@@ -1,6 +1,7 @@
 import { Inject, Injectable } from '@nestjs/common';
 import { CLIPS_REPOSITORY } from '../../domain/clips.repository';
 import type { ClipsRepository } from '../../domain/clips.repository';
+import { LikeClipOutput } from '../dtos/like-clip-output.dto';
 import { ClipsError } from '../errors/clips.error';
 
 @Injectable()
@@ -10,7 +11,7 @@ export class UnlikeClipUseCase {
     private readonly clipsRepository: ClipsRepository,
   ) {}
 
-  async execute(userId: string, clipId: string) {
+  async execute(userId: string, clipId: string): Promise<LikeClipOutput> {
     const clip = await this.clipsRepository.findClipByIdForUser(userId, clipId);
 
     if (!clip) {
