+++ src/clips/application/usecases/update-clip.usecase.ts
@@ -1,14 +1,10 @@
 import { Inject, Injectable } from '@nestjs/common';
 import { CLIPS_REPOSITORY } from '../../domain/clips.repository';
+import type { Clip } from '../../domain/clip.types';
 import type { ClipsRepository } from '../../domain/clips.repository';
+import { UpdateClipInput } from '../dtos/update-clip-input.dto';
 import { ClipsError } from '../errors/clips.error';
 import { resolveClipData } from '../policies/clip-data.policy';
-
-export type UpdateClipInput = {
-  clipId: string;
-  folderId?: string;
-  text?: string;
-};
 
 @Injectable()
 export class UpdateClipUseCase {
@@ -21,7 +17,7 @@ export class UpdateClipUseCase {
     userId: string,
     input: UpdateClipInput,
     file?: Express.Multer.File,
-  ) {
+  ): Promise<Clip> {
     const clip = await this.clipsRepository.findClipByIdForUser(
       userId,
       input.clipId,
